Rename files, move converters
